TravelAssistant/map_service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.0.0
httpx==0.24.1 